+++ app/newHome.tsx
@@ -1,0 +1,5 @@
+import ServiceStopNotification from "@/components/top/ServiceStopNotification";
+
+export default function NewHome() {
+  return <ServiceStopNotification />;
+}
