+++ lib/time-check.ts
@@ -1,0 +1,5 @@
+export function isAfterSwitchTime(): boolean {
+  const now = new Date();
+  const switchTime = new Date("2025-07-20T00:00:00+09:00");
+  return now >= switchTime;
+}
